// ceirpanel/apps/ceirpanel/src/app/core/models/filter.model.ts
export interface FilterModel {
    startDate: string;
    endDate: string;
    groupName: string;
    parentGroupName: string;
    moduleTagName: string;
    moduleName: string;
    featureName: string;
    featureModuleName: string;
    roleName: string;
    userName: string;
    firstName: string;
    lastName: string;
    organization: string;
    mobileNumber: string;
    ticketId: number;
    raisedBy: string;
    ticketStatus: string;
    companyName: string;
    userId: string;
}